+++ src/main/java/Instance/InstanceRequest.java
@@ -7,9 +7,9 @@ public class InstanceRequest {
     private int day;
     private int timeslot;
     private int gain;
-    private float penalty_A;
-    private float penalty_D;
-    private float penalty_T;
+    private double penalty_A;
+    private double penalty_D;
+    private double penalty_T;
     private int proxy;
 
 
@@ -83,27 +83,27 @@ public class InstanceRequest {
         this.gain = gain;
     }
 
-    public float getPenalty_A() {
+    public double getPenalty_A() {
         return penalty_A;
     }
 
-    public void setPenalty_A(float penalty_A) {
+    public void setPenalty_A(double penalty_A) {
         this.penalty_A = penalty_A;
     }
 
-    public float getPenalty_D() {
+    public double getPenalty_D() {
         return penalty_D;
     }
 
-    public void setPenalty_D(float penalty_D) {
+    public void setPenalty_D(double penalty_D) {
         this.penalty_D = penalty_D;
     }
 
-    public float getPenalty_T() {
+    public double getPenalty_T() {
         return penalty_T;
     }
 
-    public void setPenalty_T(float penalty_T) {
+    public void setPenalty_T(double penalty_T) {
         this.penalty_T = penalty_T;
     }
