+++ src/main/kotlin/alns/heuristics/removal/BestRatioRemoval.kt
@@ -17,8 +17,8 @@ class BestRatioRemoval : RemovalHeuristic {
         return orderedTaken.subList(0, min(orderedTaken.size, q))
     }
 
-    operator fun Boolean.times(penaltyA: Float): Float {
-        return if (this) penaltyA else 0.toFloat()
+    operator fun Boolean.times(penaltyA: Double): Double {
+        return if (this) penaltyA else 0.toDouble()
     }
 
 }
